core/tools/requirements.txt
# Pinned dependencies for uiao-core tooling (sync_canon.py, drift_detector.py, metadata_validator.py, ...)
# Keep versions tight for deterministic CI; review pins quarterly.

jsonschema==4.23.0
PyYAML==6.0.2
